+++ sdk/containerservice/arm-containerservicefleet/src/containerServiceFleetClient.ts
@@ -11,22 +11,24 @@ import * as coreRestPipeline from "@azure/core-rest-pipeline";
 import {
   PipelineRequest,
   PipelineResponse,
-  SendRequest
+  SendRequest,
 } from "@azure/core-rest-pipeline";
 import * as coreAuth from "@azure/core-auth";
 import {
   OperationsImpl,
   FleetsImpl,
+  AutoUpgradeProfilesImpl,
   FleetMembersImpl,
   UpdateRunsImpl,
-  FleetUpdateStrategiesImpl
+  FleetUpdateStrategiesImpl,
 } from "./operations";
 import {
   Operations,
   Fleets,
+  AutoUpgradeProfiles,
   FleetMembers,
   UpdateRuns,
-  FleetUpdateStrategies
+  FleetUpdateStrategies,
 } from "./operationsInterfaces";
 import { ContainerServiceFleetClientOptionalParams } from "./models";
 
@@ -44,7 +46,7 @@ export class ContainerServiceFleetClient extends coreClient.ServiceClient {
   constructor(
     credentials: coreAuth.TokenCredential,
     subscriptionId: string,
-    options?: ContainerServiceFleetClientOptionalParams
+    options?: ContainerServiceFleetClientOptionalParams,
   ) {
     if (credentials === undefined) {
       throw new Error("'credentials' cannot be null");
@@ -59,10 +61,10 @@ export class ContainerServiceFleetClient extends coreClient.ServiceClient {
     }
     const defaults: ContainerServiceFleetClientOptionalParams = {
       requestContentType: "application/json; charset=utf-8",
-      credential: credentials
+      credential: credentials,
     };
 
-    const packageDetails = `azsdk-js-arm-containerservicefleet/1.0.1`;
+    const packageDetails = `azsdk-js-arm-containerservicefleet/1.1.0-beta.1`;
     const userAgentPrefix =
       options.userAgentOptions && options.userAgentOptions.userAgentPrefix
         ? `${options.userAgentOptions.userAgentPrefix} ${packageDetails}`
@@ -72,20 +74,21 @@ export class ContainerServiceFleetClient extends coreClient.ServiceClient {
       ...defaults,
       ...options,
       userAgentOptions: {
-        userAgentPrefix
+        userAgentPrefix,
       },
       endpoint:
-        options.endpoint ?? options.baseUri ?? "https://management.azure.com"
+        options.endpoint ?? options.baseUri ?? "https://management.azure.com",
     };
     super(optionsWithDefaults);
 
     let bearerTokenAuthenticationPolicyFound: boolean = false;
     if (options?.pipeline && options.pipeline.getOrderedPolicies().length > 0) {
-      const pipelinePolicies: coreRestPipeline.PipelinePolicy[] = options.pipeline.getOrderedPolicies();
+      const pipelinePolicies: coreRestPipeline.PipelinePolicy[] =
+        options.pipeline.getOrderedPolicies();
       bearerTokenAuthenticationPolicyFound = pipelinePolicies.some(
         (pipelinePolicy) =>
           pipelinePolicy.name ===
-          coreRestPipeline.bearerTokenAuthenticationPolicyName
+          coreRestPipeline.bearerTokenAuthenticationPolicyName,
       );
     }
     if (
@@ -95,7 +98,7 @@ export class ContainerServiceFleetClient extends coreClient.ServiceClient {
       !bearerTokenAuthenticationPolicyFound
     ) {
       this.pipeline.removePolicy({
-        name: coreRestPipeline.bearerTokenAuthenticationPolicyName
+        name: coreRestPipeline.bearerTokenAuthenticationPolicyName,
       });
       this.pipeline.addPolicy(
         coreRestPipeline.bearerTokenAuthenticationPolicy({
@@ -105,9 +108,9 @@ export class ContainerServiceFleetClient extends coreClient.ServiceClient {
             `${optionsWithDefaults.endpoint}/.default`,
           challengeCallbacks: {
             authorizeRequestOnChallenge:
-              coreClient.authorizeRequestOnClaimChallenge
-          }
-        })
+              coreClient.authorizeRequestOnClaimChallenge,
+          },
+        }),
       );
     }
     // Parameter assignments
@@ -115,9 +118,10 @@ export class ContainerServiceFleetClient extends coreClient.ServiceClient {
 
     // Assigning values to Constant parameters
     this.$host = options.$host || "https://management.azure.com";
-    this.apiVersion = options.apiVersion || "2023-10-15";
+    this.apiVersion = options.apiVersion || "2024-05-02-preview";
     this.operations = new OperationsImpl(this);
     this.fleets = new FleetsImpl(this);
+    this.autoUpgradeProfiles = new AutoUpgradeProfilesImpl(this);
     this.fleetMembers = new FleetMembersImpl(this);
     this.updateRuns = new UpdateRunsImpl(this);
     this.fleetUpdateStrategies = new FleetUpdateStrategiesImpl(this);
@@ -133,7 +137,7 @@ export class ContainerServiceFleetClient extends coreClient.ServiceClient {
       name: "CustomApiVersionPolicy",
       async sendRequest(
         request: PipelineRequest,
-        next: SendRequest
+        next: SendRequest,
       ): Promise<PipelineResponse> {
         const param = request.url.split("?");
         if (param.length > 1) {
@@ -147,13 +151,14 @@ export class ContainerServiceFleetClient extends coreClient.ServiceClient {
           request.url = param[0] + "?" + newParams.join("&");
         }
         return next(request);
-      }
+      },
     };
     this.pipeline.addPolicy(apiVersionPolicy);
   }
 
   operations: Operations;
   fleets: Fleets;
+  autoUpgradeProfiles: AutoUpgradeProfiles;
   fleetMembers: FleetMembers;
   updateRuns: UpdateRuns;
   fleetUpdateStrategies: FleetUpdateStrategies;