+++ sdk/containerservice/arm-containerservicefleet/src/models/index.ts
@@ -129,6 +129,49 @@ export interface FleetListResult {
   nextLink?: string;
 }
 
+/** The FleetHubProfile configures the fleet hub. */
+export interface FleetHubProfile {
+  /** DNS prefix used to create the FQDN for the Fleet hub. */
+  dnsPrefix?: string;
+  /** The access profile for the Fleet hub API server. */
+  apiServerAccessProfile?: APIServerAccessProfile;
+  /** The agent profile for the Fleet hub. */
+  agentProfile?: AgentProfile;
+  /**
+   * The FQDN of the Fleet hub.
+   * NOTE: This property will not be serialized. It can only be populated by the server.
+   */
+  readonly fqdn?: string;
+  /**
+   * The Kubernetes version of the Fleet hub.
+   * NOTE: This property will not be serialized. It can only be populated by the server.
+   */
+  readonly kubernetesVersion?: string;
+  /**
+   * The Azure Portal FQDN of the Fleet hub.
+   * NOTE: This property will not be serialized. It can only be populated by the server.
+   */
+  readonly portalFqdn?: string;
+}
+
+/** Access profile for the Fleet hub API server. */
+export interface APIServerAccessProfile {
+  /** Whether to create the Fleet hub as a private cluster or not. */
+  enablePrivateCluster?: boolean;
+  /** Whether to enable apiserver vnet integration for the Fleet hub or not. */
+  enableVnetIntegration?: boolean;
+  /** The subnet to be used when apiserver vnet integration is enabled. It is required when creating a new Fleet with BYO vnet. */
+  subnetId?: string;
+}
+
+/** Agent profile for the Fleet hub. */
+export interface AgentProfile {
+  /** The ID of the subnet which the Fleet hub node will join on startup. If this is not specified, a vnet and subnet will be generated and used. */
+  subnetId?: string;
+  /** The virtual machine size of the Fleet hub. */
+  vmSize?: string;
+}
+
 /** Managed service identity (system assigned and/or user assigned identities) */
 export interface ManagedServiceIdentity {
   /**
@@ -207,6 +250,20 @@ export interface FleetPatch {
   tags?: { [propertyName: string]: string };
   /** Managed identity. */
   identity?: ManagedServiceIdentity;
+}
+
+/** The response of a AutoUpgradeProfile list operation. */
+export interface AutoUpgradeProfileListResult {
+  /** The AutoUpgradeProfile items on this page */
+  value: AutoUpgradeProfile[];
+  /** The link to the next page of items */
+  nextLink?: string;
+}
+
+/** The node image upgrade to be applied to the target nodes in update run. */
+export interface NodeImageSelection {
+  /** The node image upgrade type. */
+  type: NodeImageSelectionType;
 }
 
 /** The Credential results response. */
@@ -305,12 +362,6 @@ export interface ManagedClusterUpgradeSpec {
   type: ManagedClusterUpgradeType;
   /** The Kubernetes version to upgrade the member clusters to. */
   kubernetesVersion?: string;
-}
-
-/** The node image upgrade to be applied to the target nodes in update run. */
-export interface NodeImageSelection {
-  /** The node image upgrade type. */
-  type: NodeImageSelectionType;
 }
 
 /** The status of a UpdateRun. */
@@ -460,6 +511,24 @@ export interface NodeImageVersion {
   readonly version?: string;
 }
 
+/** The properties of a skip operation containing multiple skip requests. */
+export interface SkipProperties {
+  /** The targets to skip. */
+  targets: SkipTarget[];
+}
+
+/** The definition of a single skip request. */
+export interface SkipTarget {
+  /** The skip target type. */
+  type: TargetType;
+  /**
+   * The skip target's name.
+   * To skip a member/group/stage, use the member/group/stage's name;
+   * Tp skip an after stage wait, use the parent stage's name.
+   */
+  name: string;
+}
+
 /** The response of a FleetUpdateStrategy list operation. */
 export interface FleetUpdateStrategyListResult {
   /** The FleetUpdateStrategy items on this page */
@@ -493,6 +562,35 @@ export interface Fleet extends TrackedResource {
    * NOTE: This property will not be serialized. It can only be populated by the server.
    */
   readonly provisioningState?: FleetProvisioningState;
+  /** The FleetHubProfile configures the Fleet's hub. */
+  hubProfile?: FleetHubProfile;
+}
+
+/** The AutoUpgradeProfile resource. */
+export interface AutoUpgradeProfile extends ProxyResource {
+  /**
+   * If eTag is provided in the response body, it may also be provided as a header per the normal etag convention.  Entity tags are used for comparing two or more entities from the same requested resource. HTTP/1.1 uses entity tags in the etag (section 14.19), If-Match (section 14.24), If-None-Match (section 14.26), and If-Range (section 14.27) header fields.
+   * NOTE: This property will not be serialized. It can only be populated by the server.
+   */
+  readonly eTag?: string;
+  /**
+   * The provisioning state of the AutoUpgradeProfile resource.
+   * NOTE: This property will not be serialized. It can only be populated by the server.
+   */
+  readonly provisioningState?: AutoUpgradeProfileProvisioningState;
+  /** The resource id of the UpdateStrategy resource to reference. If not specified, the auto upgrade will run on all clusters which are members of the fleet. */
+  updateStrategyId?: string;
+  /** Configures how auto-upgrade will be run. */
+  channel?: UpgradeChannel;
+  /** The node image upgrade to be applied to the target clusters in auto upgrade. */
+  nodeImageSelection?: NodeImageSelection;
+  /**
+   * If set to False: the auto upgrade has effect - target managed clusters will be upgraded on schedule.
+   * If set to True: the auto upgrade has no effect - no upgrade will be run on the target managed clusters.
+   * This is a boolean and not an enum because enabled/disabled are all available states of the auto upgrade profile.
+   * By default, this is set to False.
+   */
+  disabled?: boolean;
 }
 
 /** A member of the Fleet. It contains a reference to an existing Kubernetes cluster on Azure. */
@@ -593,6 +691,20 @@ export interface FleetsDeleteHeaders {
   location?: string;
 }
 
+/** Defines headers for AutoUpgradeProfiles_createOrUpdate operation. */
+export interface AutoUpgradeProfilesCreateOrUpdateHeaders {
+  /** The Retry-After header can indicate how long the client should wait before polling the operation status. */
+  retryAfter?: number;
+}
+
+/** Defines headers for AutoUpgradeProfiles_delete operation. */
+export interface AutoUpgradeProfilesDeleteHeaders {
+  /** The Retry-After header can indicate how long the client should wait before polling the operation status. */
+  retryAfter?: number;
+  /** The Location header contains the URL where the status of the long running operation can be checked. */
+  location?: string;
+}
+
 /** Defines headers for FleetMembers_create operation. */
 export interface FleetMembersCreateHeaders {
   /** The Retry-After header can indicate how long the client should wait before polling the operation status. */
@@ -666,7 +778,7 @@ export enum KnownOrigin {
   /** System */
   System = "system",
   /** UserSystem */
-  UserSystem = "user,system"
+  UserSystem = "user,system",
 }
 
 /**
@@ -683,7 +795,7 @@ export type Origin = string;
 /** Known values of {@link ActionType} that the service accepts. */
 export enum KnownActionType {
   /** Internal */
-  Internal = "Internal"
+  Internal = "Internal",
 }
 
 /**
@@ -708,7 +820,7 @@ export enum KnownFleetProvisioningState {
   /** The provisioning state of a fleet being updated. */
   Updating = "Updating",
   /** The provisioning state of a fleet being deleted. */
-  Deleting = "Deleting"
+  Deleting = "Deleting",
 }
 
 /**
@@ -734,7 +846,7 @@ export enum KnownManagedServiceIdentityType {
   /** UserAssigned */
   UserAssigned = "UserAssigned",
   /** SystemAssignedUserAssigned */
-  SystemAssignedUserAssigned = "SystemAssigned, UserAssigned"
+  SystemAssignedUserAssigned = "SystemAssigned, UserAssigned",
 }
 
 /**
@@ -758,7 +870,7 @@ export enum KnownCreatedByType {
   /** ManagedIdentity */
   ManagedIdentity = "ManagedIdentity",
   /** Key */
-  Key = "Key"
+  Key = "Key",
 }
 
 /**
@@ -773,6 +885,70 @@ export enum KnownCreatedByType {
  */
 export type CreatedByType = string;
 
+/** Known values of {@link AutoUpgradeProfileProvisioningState} that the service accepts. */
+export enum KnownAutoUpgradeProfileProvisioningState {
+  /** Resource has been created. */
+  Succeeded = "Succeeded",
+  /** Resource creation failed. */
+  Failed = "Failed",
+  /** Resource creation was canceled. */
+  Canceled = "Canceled",
+}
+
+/**
+ * Defines values for AutoUpgradeProfileProvisioningState. \
+ * {@link KnownAutoUpgradeProfileProvisioningState} can be used interchangeably with AutoUpgradeProfileProvisioningState,
+ *  this enum contains the known values that the service supports.
+ * ### Known values supported by the service
+ * **Succeeded**: Resource has been created. \
+ * **Failed**: Resource creation failed. \
+ * **Canceled**: Resource creation was canceled.
+ */
+export type AutoUpgradeProfileProvisioningState = string;
+
+/** Known values of {@link UpgradeChannel} that the service accepts. */
+export enum KnownUpgradeChannel {
+  /**
+   *  Upgrades the clusters kubernetes version to the latest supported patch release on minor version N-1, where N is the latest supported minor version.
+   *  For example, if a cluster runs version 1.17.7 and versions 1.17.9, 1.18.4, 1.18.6, and 1.19.1 are available, the cluster upgrades to 1.18.6.
+   */
+  Stable = "Stable",
+  /** Upgrades the clusters kubernetes version to the latest supported patch release on the latest supported minor version. */
+  Rapid = "Rapid",
+  /** Upgrade node image version of the clusters. */
+  NodeImage = "NodeImage",
+}
+
+/**
+ * Defines values for UpgradeChannel. \
+ * {@link KnownUpgradeChannel} can be used interchangeably with UpgradeChannel,
+ *  this enum contains the known values that the service supports.
+ * ### Known values supported by the service
+ * **Stable**:  Upgrades the clusters kubernetes version to the latest supported patch release on minor version N-1, where N is the latest supported minor version.
+ *  For example, if a cluster runs version 1.17.7 and versions 1.17.9, 1.18.4, 1.18.6, and 1.19.1 are available, the cluster upgrades to 1.18.6. \
+ * **Rapid**: Upgrades the clusters kubernetes version to the latest supported patch release on the latest supported minor version. \
+ * **NodeImage**: Upgrade node image version of the clusters.
+ */
+export type UpgradeChannel = string;
+
+/** Known values of {@link NodeImageSelectionType} that the service accepts. */
+export enum KnownNodeImageSelectionType {
+  /** Use the latest image version when upgrading nodes. Clusters may use different image versions (e.g., 'AKSUbuntu-1804gen2containerd-2021.10.12' and 'AKSUbuntu-1804gen2containerd-2021.10.19') because, for example, the latest available version is different in different regions. */
+  Latest = "Latest",
+  /** The image versions to upgrade nodes to are selected as described below: for each node pool in managed clusters affected by the update run, the system selects the latest image version such that it is available across all other node pools (in all other clusters) of the same image type. As a result, all node pools of the same image type will be upgraded to the same image version. For example, if the latest image version for image type 'AKSUbuntu-1804gen2containerd' is 'AKSUbuntu-1804gen2containerd-2021.10.12' for a node pool in cluster A in region X, and is 'AKSUbuntu-1804gen2containerd-2021.10.17' for a node pool in cluster B in region Y, the system will upgrade both node pools to image version 'AKSUbuntu-1804gen2containerd-2021.10.12'. */
+  Consistent = "Consistent",
+}
+
+/**
+ * Defines values for NodeImageSelectionType. \
+ * {@link KnownNodeImageSelectionType} can be used interchangeably with NodeImageSelectionType,
+ *  this enum contains the known values that the service supports.
+ * ### Known values supported by the service
+ * **Latest**: Use the latest image version when upgrading nodes. Clusters may use different image versions (e.g., 'AKSUbuntu-1804gen2containerd-2021.10.12' and 'AKSUbuntu-1804gen2containerd-2021.10.19') because, for example, the latest available version is different in different regions. \
+ * **Consistent**: The image versions to upgrade nodes to are selected as described below: for each node pool in managed clusters affected by the update run, the system selects the latest image version such that it is available across all other node pools (in all other clusters) of the same image type. As a result, all node pools of the same image type will be upgraded to the same image version. For example, if the latest image version for image type 'AKSUbuntu-1804gen2containerd' is 'AKSUbuntu-1804gen2containerd-2021.10.12' for a node pool in cluster A in region X, and is 'AKSUbuntu-1804gen2containerd-2021.10.17' for a node pool in cluster B in region Y, the system will upgrade both node pools to image version 'AKSUbuntu-1804gen2containerd-2021.10.12'.
+ */
+export type NodeImageSelectionType = string;
+
 /** Known values of {@link FleetMemberProvisioningState} that the service accepts. */
 export enum KnownFleetMemberProvisioningState {
   /** Resource has been created. */
@@ -786,7 +962,7 @@ export enum KnownFleetMemberProvisioningState {
   /** The provisioning state of a member leaving a fleet. */
   Leaving = "Leaving",
   /** The provisioning state of a member being updated. */
-  Updating = "Updating"
+  Updating = "Updating",
 }
 
 /**
@@ -810,7 +986,7 @@ export enum KnownUpdateRunProvisioningState {
   /** Resource creation failed. */
   Failed = "Failed",
   /** Resource creation was canceled. */
-  Canceled = "Canceled"
+  Canceled = "Canceled",
 }
 
 /**
@@ -829,7 +1005,7 @@ export enum KnownManagedClusterUpgradeType {
   /** Full upgrades the control plane and all agent pools of the target ManagedClusters. */
   Full = "Full",
   /** NodeImageOnly upgrades only the node images of the target ManagedClusters. */
-  NodeImageOnly = "NodeImageOnly"
+  NodeImageOnly = "NodeImageOnly",
 }
 
 /**
@@ -841,24 +1017,6 @@ export enum KnownManagedClusterUpgradeType {
  * **NodeImageOnly**: NodeImageOnly upgrades only the node images of the target ManagedClusters.
  */
 export type ManagedClusterUpgradeType = string;
-
-/** Known values of {@link NodeImageSelectionType} that the service accepts. */
-export enum KnownNodeImageSelectionType {
-  /** Use the latest image version when upgrading nodes. Clusters may use different image versions (e.g., 'AKSUbuntu-1804gen2containerd-2021.10.12' and 'AKSUbuntu-1804gen2containerd-2021.10.19') because, for example, the latest available version is different in different regions. */
-  Latest = "Latest",
-  /** The image versions to upgrade nodes to are selected as described below: for each node pool in managed clusters affected by the update run, the system selects the latest image version such that it is available across all other node pools (in all other clusters) of the same image type. As a result, all node pools of the same image type will be upgraded to the same image version. For example, if the latest image version for image type 'AKSUbuntu-1804gen2containerd' is 'AKSUbuntu-1804gen2containerd-2021.10.12' for a node pool in cluster A in region X, and is 'AKSUbuntu-1804gen2containerd-2021.10.17' for a node pool in cluster B in region Y, the system will upgrade both node pools to image version 'AKSUbuntu-1804gen2containerd-2021.10.12'. */
-  Consistent = "Consistent"
-}
-
-/**
- * Defines values for NodeImageSelectionType. \
- * {@link KnownNodeImageSelectionType} can be used interchangeably with NodeImageSelectionType,
- *  this enum contains the known values that the service supports.
- * ### Known values supported by the service
- * **Latest**: Use the latest image version when upgrading nodes. Clusters may use different image versions (e.g., 'AKSUbuntu-1804gen2containerd-2021.10.12' and 'AKSUbuntu-1804gen2containerd-2021.10.19') because, for example, the latest available version is different in different regions. \
- * **Consistent**: The image versions to upgrade nodes to are selected as described below: for each node pool in managed clusters affected by the update run, the system selects the latest image version such that it is available across all other node pools (in all other clusters) of the same image type. As a result, all node pools of the same image type will be upgraded to the same image version. For example, if the latest image version for image type 'AKSUbuntu-1804gen2containerd' is 'AKSUbuntu-1804gen2containerd-2021.10.12' for a node pool in cluster A in region X, and is 'AKSUbuntu-1804gen2containerd-2021.10.17' for a node pool in cluster B in region Y, the system will upgrade both node pools to image version 'AKSUbuntu-1804gen2containerd-2021.10.12'.
- */
-export type NodeImageSelectionType = string;
 
 /** Known values of {@link UpdateState} that the service accepts. */
 export enum KnownUpdateState {
@@ -875,7 +1033,7 @@ export enum KnownUpdateState {
   /** The state of an UpdateRun\/UpdateStage\/UpdateGroup\/MemberUpdate that has failed. */
   Failed = "Failed",
   /** The state of an UpdateRun\/UpdateStage\/UpdateGroup\/MemberUpdate that has completed. */
-  Completed = "Completed"
+  Completed = "Completed",
 }
 
 /**
@@ -893,6 +1051,30 @@ export enum KnownUpdateState {
  */
 export type UpdateState = string;
 
+/** Known values of {@link TargetType} that the service accepts. */
+export enum KnownTargetType {
+  /** Skip the update of a member. */
+  Member = "Member",
+  /** Skip the update of a group. */
+  Group = "Group",
+  /** Skip the update of an entire stage including the after stage wait. */
+  Stage = "Stage",
+  /** Skip the update of the after stage wait of a certain stage. */
+  AfterStageWait = "AfterStageWait",
+}
+
+/**
+ * Defines values for TargetType. \
+ * {@link KnownTargetType} can be used interchangeably with TargetType,
+ *  this enum contains the known values that the service supports.
+ * ### Known values supported by the service
+ * **Member**: Skip the update of a member. \
+ * **Group**: Skip the update of a group. \
+ * **Stage**: Skip the update of an entire stage including the after stage wait. \
+ * **AfterStageWait**: Skip the update of the after stage wait of a certain stage.
+ */
+export type TargetType = string;
+
 /** Known values of {@link FleetUpdateStrategyProvisioningState} that the service accepts. */
 export enum KnownFleetUpdateStrategyProvisioningState {
   /** Resource has been created. */
@@ -900,7 +1082,7 @@ export enum KnownFleetUpdateStrategyProvisioningState {
   /** Resource creation failed. */
   Failed = "Failed",
   /** Resource creation was canceled. */
-  Canceled = "Canceled"
+  Canceled = "Canceled",
 }
 
 /**
@@ -1011,6 +1193,56 @@ export interface FleetsListByResourceGroupNextOptionalParams
 export type FleetsListByResourceGroupNextResponse = FleetListResult;
 
 /** Optional parameters. */
+export interface AutoUpgradeProfilesListByFleetOptionalParams
+  extends coreClient.OperationOptions {}
+
+/** Contains response data for the listByFleet operation. */
+export type AutoUpgradeProfilesListByFleetResponse =
+  AutoUpgradeProfileListResult;
+
+/** Optional parameters. */
+export interface AutoUpgradeProfilesGetOptionalParams
+  extends coreClient.OperationOptions {}
+
+/** Contains response data for the get operation. */
+export type AutoUpgradeProfilesGetResponse = AutoUpgradeProfile;
+
+/** Optional parameters. */
+export interface AutoUpgradeProfilesCreateOrUpdateOptionalParams
+  extends coreClient.OperationOptions {
+  /** The request should only proceed if an entity matches this string. */
+  ifMatch?: string;
+  /** The request should only proceed if no entity matches this string. */
+  ifNoneMatch?: string;
+  /** Delay to wait until next poll, in milliseconds. */
+  updateIntervalInMs?: number;
+  /** A serialized poller which can be used to resume an existing paused Long-Running-Operation. */
+  resumeFrom?: string;
+}
+
+/** Contains response data for the createOrUpdate operation. */
+export type AutoUpgradeProfilesCreateOrUpdateResponse = AutoUpgradeProfile;
+
+/** Optional parameters. */
+export interface AutoUpgradeProfilesDeleteOptionalParams
+  extends coreClient.OperationOptions {
+  /** The request should only proceed if an entity matches this string. */
+  ifMatch?: string;
+  /** Delay to wait until next poll, in milliseconds. */
+  updateIntervalInMs?: number;
+  /** A serialized poller which can be used to resume an existing paused Long-Running-Operation. */
+  resumeFrom?: string;
+}
+
+/** Optional parameters. */
+export interface AutoUpgradeProfilesListByFleetNextOptionalParams
+  extends coreClient.OperationOptions {}
+
+/** Contains response data for the listByFleetNext operation. */
+export type AutoUpgradeProfilesListByFleetNextResponse =
+  AutoUpgradeProfileListResult;
+
+/** Optional parameters. */
 export interface FleetMembersListByFleetOptionalParams
   extends coreClient.OperationOptions {}
 
@@ -1114,6 +1346,16 @@ export interface UpdateRunsDeleteOptionalParams
 }
 
 /** Optional parameters. */
+export interface UpdateRunsSkipOptionalParams
+  extends coreClient.OperationOptions {
+  /** The request should only proceed if an entity matches this string. */
+  ifMatch?: string;
+}
+
+/** Contains response data for the skip operation. */
+export type UpdateRunsSkipResponse = UpdateRun;
+
+/** Optional parameters. */
 export interface UpdateRunsStartOptionalParams
   extends coreClient.OperationOptions {
   /** The request should only proceed if an entity matches this string. */
@@ -1153,7 +1395,8 @@ export interface FleetUpdateStrategiesListByFleetOptionalParams
   extends coreClient.OperationOptions {}
 
 /** Contains response data for the listByFleet operation. */
-export type FleetUpdateStrategiesListByFleetResponse = FleetUpdateStrategyListResult;
+export type FleetUpdateStrategiesListByFleetResponse =
+  FleetUpdateStrategyListResult;
 
 /** Optional parameters. */
 export interface FleetUpdateStrategiesGetOptionalParams
@@ -1194,7 +1437,8 @@ export interface FleetUpdateStrategiesListByFleetNextOptionalParams
   extends coreClient.OperationOptions {}
 
 /** Contains response data for the listByFleetNext operation. */
-export type FleetUpdateStrategiesListByFleetNextResponse = FleetUpdateStrategyListResult;
+export type FleetUpdateStrategiesListByFleetNextResponse =
+  FleetUpdateStrategyListResult;
 
 /** Optional parameters. */
 export interface ContainerServiceFleetClientOptionalParams