+++ sdk/containerservice/arm-containerservicefleet/CHANGELOG.md
@@ -1,15 +1,40 @@
 # Release History
+    
+## 1.1.0-beta.1 (2024-03-28)
+    
+**Features**
 
-## 1.0.1 (Unreleased)
-
-### Features Added
-
-### Breaking Changes
-
-### Bugs Fixed
-
-### Other Changes
-
+  - Added operation group AutoUpgradeProfiles
+  - Added operation UpdateRuns.skip
+  - Added Interface AgentProfile
+  - Added Interface APIServerAccessProfile
+  - Added Interface AutoUpgradeProfile
+  - Added Interface AutoUpgradeProfileListResult
+  - Added Interface AutoUpgradeProfilesCreateOrUpdateHeaders
+  - Added Interface AutoUpgradeProfilesCreateOrUpdateOptionalParams
+  - Added Interface AutoUpgradeProfilesDeleteHeaders
+  - Added Interface AutoUpgradeProfilesDeleteOptionalParams
+  - Added Interface AutoUpgradeProfilesGetOptionalParams
+  - Added Interface AutoUpgradeProfilesListByFleetNextOptionalParams
+  - Added Interface AutoUpgradeProfilesListByFleetOptionalParams
+  - Added Interface FleetHubProfile
+  - Added Interface SkipProperties
+  - Added Interface SkipTarget
+  - Added Interface UpdateRunsSkipOptionalParams
+  - Added Type Alias AutoUpgradeProfileProvisioningState
+  - Added Type Alias AutoUpgradeProfilesCreateOrUpdateResponse
+  - Added Type Alias AutoUpgradeProfilesGetResponse
+  - Added Type Alias AutoUpgradeProfilesListByFleetNextResponse
+  - Added Type Alias AutoUpgradeProfilesListByFleetResponse
+  - Added Type Alias TargetType
+  - Added Type Alias UpdateRunsSkipResponse
+  - Added Type Alias UpgradeChannel
+  - Interface Fleet has a new optional parameter hubProfile
+  - Added Enum KnownAutoUpgradeProfileProvisioningState
+  - Added Enum KnownTargetType
+  - Added Enum KnownUpgradeChannel
+    
+    
 ## 1.0.0 (2023-10-27)
 
 The package of @azure/arm-containerservicefleet is using our next generation design principles. To learn more, please refer to our documentation [Quick Start](https://aka.ms/azsdk/js/mgmt/quickstart).